Fix scraper and add scheduler (#4)
* Fix scraper and add scheduler

The scraper was fixed because it was only retrieving one news per site. The scheduler will be added in this PR as well because of its ease to use.

* Add scheduler and fix classification

The scheduler has been added, it should be scraping news at 03:00 every day. There was a problem classifying, the compare operators returned an object instead of a boolean.

* Refactor classifier

The classifier instance was put inside a function so that the code becomes easier to read.

* Remove comments